test-suite/requirements.txt
apache-airflow==2.3.3
apache-airflow-providers-amazon==6.1.0
openmetadata-ingestion~=0.12.3.0
pandas==1.3.5
requests==2.28.1